% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center metric data matrix by column means.}
\usage{
center(X)
}
\arguments{
\item{X}{A complete (N x n) metric data matrix of N obs units and n features.}
}
\value{
Centered data matrix.
}
\description{
Center metric data matrix by column means.
}
