% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Implementation of the Grassberger-Procaccia algorithm}
\usage{
gp(D, r = NULL)
}
\arguments{
\item{D}{A symmetric distance / dissimilarity matrix of N objects}

\item{r}{Optional positive vector of distance cutoffs at which the correlation integral is approximated}
}
\value{
List of results
}
\description{
Implementation of the Grassberger-Procaccia algorithm
}
