% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp_graph_analyses}
\alias{gp_graph_analyses}
\title{Graphical analyses of Grassberger-Procaccia approximated cor integral}
\usage{
gp_graph_analyses(gp.result, ...)
}
\arguments{
\item{gp.result}{A list of results obtained through the gp function}

\item{...}{Options to be passed to plot}
}
\value{
List of results.
}
\description{
Graphical analyses of Grassberger-Procaccia approximated cor integral
}
