% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanspp_init.R
\name{kmpp_init}
\alias{kmpp_init}
\title{Simple implementation of k-means++ initialized prototypes. Utilizes sqrd. euclidean distance!}
\usage{
kmpp_init(X, k = 3)
}
\arguments{
\item{X}{The (N x n) data matrix}

\item{k}{Number of prototypes}
}
\value{
A (K x n) matrix of prototypes
}
\description{
Simple implementation of k-means++ initialized prototypes. Utilizes sqrd. euclidean distance!
}
