% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_gas.R
\name{neural_gas}
\alias{neural_gas}
\title{Implementation of the Neural Gas algorithm}
\usage{
neural_gas(
  X,
  K = ceiling(nrow(X)/20),
  W = NULL,
  lambda_init = NULL,
  lambda_stop = 0.001,
  learn_rate = 0.1,
  max_iter = nrow(X) * 100,
  verbose = TRUE
)
}
\arguments{
\item{X}{Design or raw data (N x n) matrix}

\item{K}{Number of prototypes to compute, defaults to ceiling(nrow(X) / 20)}

\item{W}{Optional (K x n) matrix of initial prototype data. If given, the K argument is ignored}

\item{lambda_init}{Initial general adaption rate}

\item{lambda_stop}{Target adaption rate during last adaption step. Default is 0.001.}

\item{learn_rate}{Constant adaption rate. Default is 0.1}

\item{max_iter}{The number of steps to perform. Defaults to N x 100.}

\item{verbose}{Should outputs be printed. Default is TRUE.}
}
\value{
(K x n) matrix of prototype data after learning
}
\description{
Implementation of the Neural Gas algorithm
}
