% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.ggplots.R
\name{pca_ggplots}
\alias{pca_ggplots}
\title{Title}
\usage{
pca_ggplots(pca_res, class = NULL, varnames = NULL, list_plots = FALSE)
}
\arguments{
\item{pca_res}{Results of PCA computations, obtained through pca function}

\item{class}{Optional indicator vector of input data size, stating some class information per data point. Used for highlighting plotted projection scores.}

\item{varnames}{Optional vector of n feature names.}

\item{list_plots}{logical. Should the ggplot objects be returned as list. Default is FALSE?}
}
\value{
If list_plots is set to TRUE, ggplots are returned as list.
}
\description{
Title
}
