% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale z-scaled data}
\usage{
rescale(
  scaled_data,
  mean = ifelse(is.null(attr(scaled_data, "scaled:center")), 0, attr(scaled_data,
    "scaled:center")),
  scale = ifelse(is.null(attr(scaled_data, "scaled:scale")), 1, attr(scaled_data,
    "scaled:scale"))
)
}
\arguments{
\item{mean}{Provide a mean for rescaling. Otherwise attribute mean data is read from the scaled_data object}

\item{scale}{Provide a standard deviation for rescaling. Otherwise attribute std data is read from the scaled_data object}
}
\description{
Rescale z-scaled data
}
