% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleN.R
\name{scaleN}
\alias{scaleN}
\title{z-scaling for some vector using "geometric" standard deviation}
\usage{
scaleN(x, center = T, scale = T)
}
\arguments{
\item{x}{The data vector}

\item{center}{logical. Invoke data centering.}

\item{scale}{logical. Invoke data scaling by factor 1/N, instead of Bessel correction.}
}
\value{
Scaled data vector.
}
\description{
z-scaling for some vector using "geometric" standard deviation
}
