% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdN.R
\name{sdN}
\alias{sdN}
\title{Compute "geometric" standard deviation, without Bessel correction.}
\usage{
sdN(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector containing metric data.}

\item{na.rm}{Logical; remove NA data prior to computations. Defaults to FALSE.}
}
\value{
Scalar value.
}
\description{
Compute "geometric" standard deviation, without Bessel correction.
}
