% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varN.R
\name{varN}
\alias{varN}
\title{Compute "geometric" variance, without Bessel correction.}
\usage{
varN(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector containing metric data.}

\item{na.rm}{Logical; remove NA data prior to computations. Defaults to FALSE.}
}
\value{
A scalar value
}
\description{
Compute "geometric" variance, without Bessel correction.
}
