% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Performs classic principal component "analysis" - or objects usable for analysis, at least}
\usage{
pca(X, center = T, scale = T, k = 2)
}
\arguments{
\item{k}{}
}
\description{
Performs classic principal component "analysis" - or objects usable for analysis, at least
}
